+++ server/src/main/scala/de/frosner/broccoli/auth/UserAccount.scala
@@ -1,4 +1,4 @@
-package de.frosner.broccoli.models
+package de.frosner.broccoli.auth
 
 import de.frosner.broccoli.RemoveSecrets
 import play.api.libs.json.{Json, Writes}
