+++ server/src/main/scala/de/frosner/broccoli/controllers/AboutController.scala
@@ -2,10 +2,10 @@ package de.frosner.broccoli.controllers
 
 import javax.inject.Inject
 
+import de.frosner.broccoli.auth.Account
 import de.frosner.broccoli.services._
 import de.frosner.broccoli.conf
 import de.frosner.broccoli.models.AboutInfo.aboutInfoWrites
-import de.frosner.broccoli.models.Account
 import jp.t2v.lab.play2.auth.BroccoliSimpleAuthorization
 import play.api.Environment
 import play.api.cache.CacheApi
