+++ server/src/main/scala/de/frosner/broccoli/websocket/WebSocketMessageHandler.scala
@@ -4,9 +4,9 @@ import javax.inject.Inject
 
 import cats.instances.future._
 import cats.syntax.either._
+import de.frosner.broccoli.auth.Account
 import de.frosner.broccoli.controllers.InstanceController
 import de.frosner.broccoli.instances.NomadInstances
-import de.frosner.broccoli.models.Account
 import de.frosner.broccoli.services.InstanceService
 import de.frosner.broccoli.websocket.IncomingMessage._
 import de.frosner.broccoli.websocket.OutgoingMessage._
