+++ server/src/test/scala/de/frosner/broccoli/controllers/InstanceControllerSpec.scala
@@ -3,6 +3,7 @@ package de.frosner.broccoli.controllers
 import cats.data.EitherT
 import cats.instances.future._
 import com.mohiva.play.silhouette.api.util.Credentials
+import de.frosner.broccoli.auth.{Account, Role, UserAccount}
 import de.frosner.broccoli.http.ToHTTPResult
 import de.frosner.broccoli.instances.NomadInstances
 import de.frosner.broccoli.instances.InstanceNotFoundException
