+++ server/src/test/scala/de/frosner/broccoli/controllers/SecurityControllerSpec.scala
@@ -3,7 +3,7 @@ package de.frosner.broccoli.controllers
 import java.util.concurrent.TimeUnit
 
 import com.mohiva.play.silhouette.api.util.Credentials
-import de.frosner.broccoli.models.{Role, UserAccount}
+import de.frosner.broccoli.auth.{Role, UserAccount}
 import de.frosner.broccoli.services.{SecurityService, WebSocketService}
 import jp.t2v.lab.play2.auth.test.Helpers._
 import org.mockito.Matchers._
