+++ server/src/test/scala/de/frosner/broccoli/models/AccountSpec.scala
@@ -3,6 +3,7 @@ package de.frosner.broccoli.models
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
 import de.frosner.broccoli.RemoveSecrets.ToRemoveSecretsOps
+import de.frosner.broccoli.auth.{Account, UserAccount}
 
 class AccountSpec extends Specification with ScalaCheck with ModelArbitraries with ToRemoveSecretsOps {
   "The RemoveSecrets instance" should {
