+++ server/src/test/scala/de/frosner/broccoli/services/WebSocketServiceSpec.scala
@@ -1,14 +1,6 @@
 package de.frosner.broccoli.services
 
-import com.google.common.collect.ImmutableMap
-import com.typesafe.config.{ConfigFactory, ConfigValueFactory}
-import de.frosner.broccoli.conf
-import de.frosner.broccoli.models.{Role, UserAccount}
 import org.specs2.mutable.Specification
-import play.api.Configuration
-
-import scala.collection.JavaConverters._
-import scala.util.Success
 
 class WebSocketServiceSpec extends Specification {
