+++ server/src/test/scala/de/frosner/broccoli/websocket/BroccoliMessageHandlerSpec.scala
@@ -2,6 +2,7 @@ package de.frosner.broccoli.websocket
 
 import cats.data.EitherT
 import cats.instances.future._
+import de.frosner.broccoli.auth.{Account, UserAccount}
 import de.frosner.broccoli.instances.NomadInstances
 import de.frosner.broccoli.models._
 import de.frosner.broccoli.nomad
@@ -23,7 +24,7 @@ class BroccoliMessageHandlerSpec
 
   "The Broccoli Message Handler" should {
     "send back instance tasks" in { implicit ee: ExecutionEnv =>
-      prop { (account: UserAccount, id: String, tasks: List[AllocatedTask]) =>
+      prop { (account: UserAgccount, id: String, tasks: List[AllocatedTask]) =>
         val instances = mock[NomadInstances]
         val instanceTasks = InstanceTasks(id, tasks)
         instances.getInstanceTasks(account)(id) returns EitherT.pure[Future, InstanceError](instanceTasks)
